Create DiscountRevenue module
