+++ app/models/merchant.rb
@@ -1,4 +1,6 @@
 class Merchant < ApplicationRecord
+  include DiscountRevenue
+
   validates_presence_of :name
   has_many :items
   has_many :invoice_items, through: :items
@@ -53,21 +55,5 @@ class Merchant < ApplicationRecord
             .group("invoices.created_at")
             .order("revenue desc", "invoices.created_at desc")
             .first&.created_at&.to_date
-  end
-
-  def total_revenue
-    invoice_items.sum("invoice_items.unit_price * invoice_items.quantity")
-  end
-
-  def discounted_amount
-    invoice_items
-    .joins(:bulk_discounts)
-    .where('invoice_items.quantity >= bulk_discounts.quantity')
-    .select('max(invoice_items.quantity * invoice_items.unit_price * bulk_discounts.percent / 100.0)')
-    .group(:id).sum(&:max)
-  end
-
-  def discounted_revenue
-    total_revenue - discounted_amount
   end
 end
